rename login component
